+++ themes/Lokaal-Hellendoorn/single-nieuws.php
@@ -22,6 +22,8 @@
           <img src="<?php the_field('nieuws_afbeelding');?>" alt="news image">
         </div>
         <div class="col-xs-12 col-sm-6">
+            <p class="post-publish-date-single">Gepubliceerd op <?php the_time( 'j F, Y' ); ?>.</p>
+
             <?php the_field('nieuws_tekst');?>
         </div>
     </div>