+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 * Styling for "Media & Text" header (Raadsleden pagina)
 * Styling for table added
 * "Fractieleden" post type removed
+* Publish date added to news post
 
 2020-05-07 - Version 1.5.3
 --------------------------